\name{print.rfsgt}
\alias{print.rfsgt}
\alias{print.vimp.rfsgt}
\title{Print Output from a Random Forest Super Greedy Tree Analysis}
\description{
  Print summary output from a Random Forest SGT analysis. This is the default
  print method for the package.
}
\usage{\method{print}{rfsgt}(x, ...)}
\arguments{
  \item{x}{An object of class \code{(rfsgt, grow)}.}
  \item{...}{Further arguments passed to or from other methods.}
}
\author{
  Hemant Ishwaran and Udaya B. Kogalur
}
\keyword{print}
